updating SevenZip
- Stop static linking of SevenZip's fixed program and adopt dynamic
linking of the official DLL.
- On other operating systems we need to use p7zip instead of SevenZip, so
we will do some additional work.

+++ Qt7z/Qt7z/7zip/CPP/7zip/UI/FileManager/ExtractCallback.cpp
@@ -759,11 +759,15 @@ STDMETHODIMP CExtractCallbackImp::AskWrite(
       destPathResultTemp = fs2us(destPathSys);
     }
     else
+    {
+      if (NFind::DoesFileExist(destPathSys))
       if (!NDir::DeleteFileAlways(destPathSys))
+      if (GetLastError() != ERROR_FILE_NOT_FOUND)
       {
         RINOK(MessageError("can not delete output file", destPathSys));
         return E_ABORT;
       }
+    }
   }
   *writeAnswer = BoolToInt(true);
   return StringToBstr(destPathResultTemp, destPathResult);
@@ -1025,7 +1029,7 @@ HRESULT CVirtFileSystem::FlushToDisk(bool closeLast)
     _numFlushed++;
     _fileIsOpen = false;
     if (file.AttribDefined)
-      NDir::SetFileAttrib(path, file.Attrib);
+      NDir::SetFileAttrib_PosixHighDetect(path, file.Attrib);
   }
   return S_OK;
 }
